+++ config/hard/Breakout/code/CGCode.java
@@ -1,0 +1,31 @@
+import java.util.*;
+import java.io.*;
+import java.math.*;
+
+class Solution {
+
+    public static void main(String args[]) {
+        Scanner in = new Scanner(System.in);
+        int bX = in.nextInt();
+        int bY = in.nextInt();
+        int vX = in.nextInt();
+        int vY = in.nextInt();
+        int pN = in.nextInt();
+        int kN = in.nextInt();
+        for (int i = 0; i < pN; i++) {
+            int pX = in.nextInt();
+            int pY = in.nextInt();
+        }
+        for (int i = 0; i < kN; i++) {
+            int kX = in.nextInt();
+            int kY = in.nextInt();
+            int kStrength = in.nextInt();
+            int kPoints = in.nextInt();
+        }
+
+        // Write an answer using System.out.println()
+        // To debug: System.err.println("Debug messages...");
+
+        System.out.println("points");
+    }
+}