+++ config/hard/Breakout/code/CGCode.js
@@ -1,0 +1,25 @@
+var inputs = readline().split(' ');
+const bX = parseInt(inputs[0]);
+const bY = parseInt(inputs[1]);
+var inputs = readline().split(' ');
+const vX = parseInt(inputs[0]);
+const vY = parseInt(inputs[1]);
+const pN = parseInt(readline());
+const kN = parseInt(readline());
+for (let i = 0; i < pN; i++) {
+    var inputs = readline().split(' ');
+    const pX = parseInt(inputs[0]);
+    const pY = parseInt(inputs[1]);
+}
+for (let i = 0; i < kN; i++) {
+    var inputs = readline().split(' ');
+    const kX = parseInt(inputs[0]);
+    const kY = parseInt(inputs[1]);
+    const kStrength = parseInt(inputs[2]);
+    const kPoints = parseInt(inputs[3]);
+}
+
+// Write an answer using console.log()
+// To debug: console.error('Debug messages...');
+
+console.log('points');