+++ config/hard/Breakout/code/CGCode.ts
@@ -1,0 +1,25 @@
+var inputs: string[] = readline().split(' ');
+const bX: number = parseInt(inputs[0]);
+const bY: number = parseInt(inputs[1]);
+var inputs: string[] = readline().split(' ');
+const vX: number = parseInt(inputs[0]);
+const vY: number = parseInt(inputs[1]);
+const pN: number = parseInt(readline());
+const kN: number = parseInt(readline());
+for (let i = 0; i < pN; i++) {
+    var inputs: string[] = readline().split(' ');
+    const pX: number = parseInt(inputs[0]);
+    const pY: number = parseInt(inputs[1]);
+}
+for (let i = 0; i < kN; i++) {
+    var inputs: string[] = readline().split(' ');
+    const kX: number = parseInt(inputs[0]);
+    const kY: number = parseInt(inputs[1]);
+    const kStrength: number = parseInt(inputs[2]);
+    const kPoints: number = parseInt(inputs[3]);
+}
+
+// Write an answer using console.log()
+// To debug: console.error('Debug messages...');
+
+console.log('points');