+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - configuration for "Alternative vote".
 - configuration for "Hanoi tower".
 - configuration for "Apple tree".
+- configuration for "Breakout".
 
 ### Fixed
 - name and link of "The Total is Right - Weird edition".